Reduce augustus tmp file memory usage
